GRAVITATIONAL_CONSTANT as double

+++ SimulationEngineLib/cpp/constant/PhysicalConstants.h
@@ -2,6 +2,6 @@
 
 class PhysicalConstants {
 public:
-    static constexpr float GRAVITATIONAL_CONSTANT = 6.6743e-11;
+    static constexpr double GRAVITATIONAL_CONSTANT = 6.6743e-11;
     static constexpr double SPEED_OF_LIGHT = 299792458.0;
 };
